chore(refactor): redo ffi packaging
